+++ empty.js
@@ -1,3 +1,0 @@
-// These functions can be stubbed out in specific environments
-export const unstable_batchedUpdates = undefined
-export const findDOMNode = undefined