+++ src/inject.js
@@ -1,6 +1,5 @@
 import React, { Component, createElement } from "react"
 import hoistStatics from "hoist-non-react-statics"
-import * as PropTypes from "./propTypes"
 import { observer } from "./observer"
 import { isStateless } from "./utils/utils"
 import { MobXProviderContext } from "./Provider"
@@ -71,7 +70,7 @@ function grabStoresByName(storeNames) {
  * or a function that manually maps the available stores from the context to props:
  * storesToProps(mobxStores, props, context) => newProps
  */
-export default function inject(/* fn(stores, nextProps) or ...storeNames */ ...storeNames) {
+export function inject(/* fn(stores, nextProps) or ...storeNames */ ...storeNames) {
     let grabStoresFn
     if (typeof arguments[0] === "function") {
         grabStoresFn = arguments[0]