+++ test/ts/compile-ts.tsx
@@ -69,35 +69,6 @@ React.createElement(observer(T7), { pizza: 4 })
 
 ReactDOM.render(<T5 />, document.body)
 
-/// with stores
-@observer(["store1", "store2"])
-class T8 extends Component<{ pizza: number }, {}> {
-    render() {
-        return <div>{this.props.pizza}</div>
-    }
-}
-
-const T9 = observer(
-    ["stores"],
-    createClass({
-        getDefaultProps() {
-            return { cake: 7 }
-        },
-        render() {
-            return (
-                <div>
-                    <T1 pizza={this.props.cake} />
-                </div>
-            )
-        }
-    })
-)
-
-const T10 = observer(["stores"], (props: { hamburger: number }) => {
-    return <T2 cake={props.hamburger} />
-})
-
-React.createElement(observer(T8), { pizza: 4 })
 
 class ProviderTest extends Component<any, any> {
     render() {
@@ -220,8 +191,6 @@ class LoginContainer2 extends Component<
 }
 
 ReactDOM.render(<LoginContainer2 x="test" />, document.body)
-
-ReactDOM.render(<T10 hamburger={3} />, document.body)
 
 class ObserverTest extends Component<any, any> {
     render() {